application/backend

+++ mrp/src/main/java/com/project/mrp/entities/Camera.kt
@@ -9,5 +9,5 @@ data class Camera(
     @field:SerializedName("full_name") val fullName: String?,
     @field:SerializedName("id") val id: Long?,
     @field:SerializedName("name") val name: String?,
-    @field:SerializedName("rover_id") val roverId: Int?
+    @field:SerializedName("rover_id") val roverId: Long?
 ) : Parcelable
